Guaranteed Minimum Lectures + better validateGameState
Each course now has a minimum amount of lectures that will appear each block.

The old validateGameState did not work well because it recursively went through everything and sometimes complained that a currency in a quest had missing fields. This is because some currencies don't have all fields, e.g. the courseIndex field.